feat: extract logic to hooks for readability
